C#: Separate visitors to dedicated files, rename and reorganize comment extraction related classes
