+++ csharp/extractor/Semmle.Extraction.CSharp/Populators/TypeContainerVisitor.cs
@@ -1,10 +1,7 @@
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
-using Semmle.Extraction.CSharp.Entities;
 using Semmle.Extraction.Entities;
-using Semmle.Util;
-using Semmle.Util.Logging;
 using System;
 using System.Collections.Generic;
 using System.IO;
@@ -80,71 +77,6 @@ namespace Semmle.Extraction.CSharp.Populators
                     cx.BindComments(ae, attribute.GetLocation());
                 }
             }
-        }
-    }
-
-    internal class TypeOrNamespaceVisitor : TypeContainerVisitor
-    {
-        public TypeOrNamespaceVisitor(Context cx, TextWriter trapFile, IEntity parent)
-            : base(cx, trapFile, parent) { }
-
-        public override void VisitUsingDirective(UsingDirectiveSyntax usingDirective)
-        {
-            // Only deal with "using namespace" not "using X = Y"
-            if (usingDirective.Alias == null)
-                new UsingDirective(cx, usingDirective, (NamespaceDeclaration)parent);
-        }
-
-        public override void VisitNamespaceDeclaration(NamespaceDeclarationSyntax node)
-        {
-            NamespaceDeclaration.Create(cx, node, (NamespaceDeclaration)parent);
-        }
-    }
-
-    internal class CompilationUnitVisitor : TypeOrNamespaceVisitor
-    {
-        public CompilationUnitVisitor(Context cx)
-            : base(cx, cx.TrapWriter.Writer, null) { }
-
-        public override void VisitExternAliasDirective(ExternAliasDirectiveSyntax node)
-        {
-            // This information is not yet extracted.
-            cx.ExtractionError("Not implemented extern alias directive", node.ToFullString(), Extraction.Entities.Location.Create(cx, node.GetLocation()), "", Severity.Info);
-        }
-
-        public override void VisitCompilationUnit(CompilationUnitSyntax compilationUnit)
-        {
-            foreach (var m in compilationUnit.ChildNodes())
-            {
-                cx.Try(m, null, () => ((CSharpSyntaxNode)m).Accept(this));
-            }
-
-            // Gather comments:
-            foreach (var trivia in compilationUnit.DescendantTrivia(compilationUnit.Span))
-            {
-                CommentLine.Extract(cx, trivia);
-            }
-
-            foreach (var trivia in compilationUnit.GetLeadingTrivia())
-            {
-                CommentLine.Extract(cx, trivia);
-            }
-
-            foreach (var trivia in compilationUnit.GetTrailingTrivia())
-            {
-                CommentLine.Extract(cx, trivia);
-            }
-        }
-    }
-
-    public class CompilationUnit
-    {
-        public static void Extract(Context cx, SyntaxNode unit)
-        {
-            // Ensure that the file itself is populated in case the source file is totally empty
-            Semmle.Extraction.Entities.File.Create(cx, unit.SyntaxTree.FilePath);
-
-            ((CSharpSyntaxNode)unit).Accept(new CompilationUnitVisitor(cx));
         }
     }
 }
