+++ csharp/extractor/Semmle.Extraction.CSharp/Populators/TypeOrNamespaceVisitor.cs
@@ -1,0 +1,24 @@
+using Microsoft.CodeAnalysis.CSharp.Syntax;
+using Semmle.Extraction.CSharp.Entities;
+using System.IO;
+
+namespace Semmle.Extraction.CSharp.Populators
+{
+    internal class TypeOrNamespaceVisitor : TypeContainerVisitor
+    {
+        public TypeOrNamespaceVisitor(Context cx, TextWriter trapFile, IEntity parent)
+            : base(cx, trapFile, parent) { }
+
+        public override void VisitUsingDirective(UsingDirectiveSyntax usingDirective)
+        {
+            // Only deal with "using namespace" not "using X = Y"
+            if (usingDirective.Alias == null)
+                new UsingDirective(cx, usingDirective, (NamespaceDeclaration)parent);
+        }
+
+        public override void VisitNamespaceDeclaration(NamespaceDeclarationSyntax node)
+        {
+            NamespaceDeclaration.Create(cx, node, (NamespaceDeclaration)parent);
+        }
+    }
+}
